warnings

+++ userspace/libsinsp/utils.cpp
@@ -1035,9 +1035,9 @@ void replace_in_place(string& str, string& substr_to_replace, string& new_substr
 ///////////////////////////////////////////////////////////////////////////////
 // sinsp_numparser implementation
 ///////////////////////////////////////////////////////////////////////////////
-uint32_t sinsp_numparser::parseu8(const string& str)
+uint8_t sinsp_numparser::parseu8(const string& str)
 {
-	uint32_t res;
+	uint8_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRIu8 "%c", &res, &temp) != 1)
@@ -1048,9 +1048,9 @@ uint32_t sinsp_numparser::parseu8(const string& str)
 	return res;
 }
 
-int32_t sinsp_numparser::parsed8(const string& str)
+int8_t sinsp_numparser::parsed8(const string& str)
 {
-	int32_t res;
+	int8_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRId8 "%c", &res, &temp) != 1)
@@ -1061,9 +1061,9 @@ int32_t sinsp_numparser::parsed8(const string& str)
 	return res;
 }
 
-uint32_t sinsp_numparser::parseu16(const string& str)
+uint16_t sinsp_numparser::parseu16(const string& str)
 {
-	uint32_t res;
+	uint16_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRIu16 "%c", &res, &temp) != 1)
@@ -1074,9 +1074,9 @@ uint32_t sinsp_numparser::parseu16(const string& str)
 	return res;
 }
 
-int32_t sinsp_numparser::parsed16(const string& str)
+int16_t sinsp_numparser::parsed16(const string& str)
 {
-	int32_t res;
+	int16_t res;
 	char temp;
 
 	if(std::sscanf(str.c_str(), "%" PRId16 "%c", &res, &temp) != 1)
